Move src/ into src/converter
In preparation for having assets stored in here.
