+++ src/converter/postprocessor/footnotes.js
